[ulp_ina219] Type optimizations to save a little space

+++ esphome/components/ulp_ina219/ulp/ina219.c
@@ -5,9 +5,9 @@
 
 static inline int32_t _ceilf(float x) { return (int32_t) (x + 0.999999f); }
 
-esp_err_t ina219_power_down(uint16_t address) { return i2c_write16(address, INA219_REGISTER_CONFIG, 0x0); }
+esp_err_t ina219_power_down(uint8_t address) { return i2c_write16(address, INA219_REGISTER_CONFIG, 0x0); }
 
-esp_err_t ina219_init(uint16_t address, float max_voltage, float r_shunt, float max_current,
+esp_err_t ina219_init(uint8_t address, float max_voltage, float r_shunt, float max_current,
                       uint32_t cfg_calibration_register, uint32_t *calibration_register, uint32_t *current_lsb) {
   esp_err_t ret = ESP_OK;
   ret = i2c_write16(address, INA219_REGISTER_CONFIG, 0x8000);
@@ -66,7 +66,7 @@ esp_err_t ina219_init(uint16_t address, float max_voltage, float r_shunt, float 
   return ret;
 }
 
-esp_err_t ina219_bus_voltage(uint16_t address, float *bus_voltage) {
+esp_err_t ina219_bus_voltage(uint8_t address, float *bus_voltage) {
   uint16_t raw_bus_voltage = 0;
   esp_err_t ret = i2c_read16(address, INA219_REGISTER_BUS_VOLTAGE, &raw_bus_voltage);
   if (ret == ESP_OK) {
@@ -76,7 +76,7 @@ esp_err_t ina219_bus_voltage(uint16_t address, float *bus_voltage) {
   return ret;
 }
 
-esp_err_t ina219_current(uint16_t address, float *current, uint32_t current_lsb) {
+esp_err_t ina219_current(uint8_t address, float *current, uint32_t current_lsb) {
   uint16_t raw_current = 0;
   esp_err_t ret = i2c_read16(address, INA219_REGISTER_CURRENT, &raw_current);
   if (ret == ESP_OK) {
@@ -86,7 +86,7 @@ esp_err_t ina219_current(uint16_t address, float *current, uint32_t current_lsb)
   return ret;
 }
 
-esp_err_t ina219_shunt_voltage(uint16_t address, float *shunt_voltage) {
+esp_err_t ina219_shunt_voltage(uint8_t address, float *shunt_voltage) {
   uint16_t raw_shunt_voltage = 0;
   esp_err_t ret = i2c_read16(address, INA219_REGISTER_SHUNT_VOLTAGE, &raw_shunt_voltage);
   if (ret == ESP_OK) {
@@ -95,7 +95,7 @@ esp_err_t ina219_shunt_voltage(uint16_t address, float *shunt_voltage) {
   return ret;
 }
 
-esp_err_t ina219_power(uint16_t address, float *power, uint32_t current_lsb) {
+esp_err_t ina219_power(uint8_t address, float *power, uint32_t current_lsb) {
   uint16_t raw_power = 0;
   esp_err_t ret = i2c_read16(address, INA219_REGISTER_POWER, &raw_power);
   if (ret == ESP_OK) {
